+++ src/types.ts
@@ -1,5 +1,6 @@
 export interface ObsidianGitSettings {
     commitMessage: string;
+    autoCommitMessage: string;
     commitDateFormat: string;
     autoSaveInterval: number;
     autoPullInterval: number;
@@ -20,6 +21,8 @@ export interface ObsidianGitSettings {
      * @deprecated Migrated to `syncMethod = 'merge'`
      */
     mergeOnPull?: boolean;
+    refreshSourceControl: boolean;
+    basePath: string;
 }
 
 export type SyncMethod = 'rebase' | 'merge' | 'reset';
@@ -63,4 +66,9 @@ export interface TreeItem {
     title: string;
     statusResult?: FileStatusResult;
     children?: TreeItem[];
+}
+
+export interface DiffViewState {
+    staged: boolean,
+    file: string,
 }